+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Use --query-as-item global flag to always add an Alfred item based on user's exact entry.
+- Alwyas show a tag that exactly matches user's input at the top of list.
 
 ## [0.16.0] - 2021-08-23
 Bumped minor version since new fuzzy search engine may produce different search results.